Move command implementation to openapi module
