+++ openapi-cli/src/main/java/io/ballerina/openapi/cmd/OpenApiMesseges.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package io.ballerina.openapi;
+package io.ballerina.openapi.cmd;
 
 import java.util.Arrays;
 import java.util.Collections;
