+++ openapi-cli/src/test/java/io/ballerina/generators/schema/AdvanceRecordTypeTests.java
@@ -23,7 +23,7 @@ import io.ballerina.compiler.syntax.tree.RecordTypeDescriptorNode;
 import io.ballerina.compiler.syntax.tree.SyntaxTree;
 import io.ballerina.compiler.syntax.tree.TypeDefinitionNode;
 import io.ballerina.generators.BallerinaSchemaGenerator;
-import io.ballerina.openapi.CodeGenerator;
+import io.ballerina.openapi.cmd.CodeGenerator;
 import io.ballerina.openapi.exception.BallerinaOpenApiException;
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.media.ObjectSchema;
